+++ core/pkg/service/arc/task/task_suite_test.go
@@ -1,0 +1,29 @@
+// Copyright 2026 Synnax Labs, Inc.
+//
+// Use of this software is governed by the Business Source License included in the file
+// licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with the Business Source
+// License, use of this software will be governed by the Apache License, Version 2.0,
+// included in the file licenses/APL.txt.
+
+package task_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+	. "github.com/synnaxlabs/x/testutil"
+)
+
+func TestArcTask(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Service Arc Task Suite")
+}
+
+var _ = BeforeSuite(func() {
+	ShouldNotLeakGoroutines()
+})
+
+var _ = ShouldNotLeakGoroutinesPerSpec()
