+++ core/pkg/service/arc/task/task_test.go
@@ -7,7 +7,7 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-package runtime_test
+package task_test
 
 import (
 	"context"
@@ -25,8 +25,8 @@ import (
 	"github.com/synnaxlabs/synnax/pkg/distribution/framer/frame"
 	"github.com/synnaxlabs/synnax/pkg/distribution/mock"
 	svcarc "github.com/synnaxlabs/synnax/pkg/service/arc"
-	"github.com/synnaxlabs/synnax/pkg/service/arc/runtime"
 	arcstatus "github.com/synnaxlabs/synnax/pkg/service/arc/status"
+	arctask "github.com/synnaxlabs/synnax/pkg/service/arc/task"
 	"github.com/synnaxlabs/synnax/pkg/service/channel"
 	"github.com/synnaxlabs/synnax/pkg/service/driver"
 	"github.com/synnaxlabs/synnax/pkg/service/label"
@@ -71,7 +71,7 @@ var _ = Describe("Task", Ordered, func() {
 	})
 
 	newFactoryWith := func(getModule func(context.Context, uuid.UUID) (svcarc.Arc, error)) driver.Factory {
-		return MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+		return MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 			Channel:    channel.Wrap(dist.Channel),
 			Framer:     dist.Framer,
 			Status:     statusSvc,
@@ -103,7 +103,7 @@ var _ = Describe("Task", Ordered, func() {
 		})
 	}
 
-	configToMap := func(cfg runtime.TaskConfig) map[string]any {
+	configToMap := func(cfg arctask.TaskConfig) map[string]any {
 		cfgJSON := MustSucceed(json.Marshal(cfg))
 		var cfgMap map[string]any
 		Expect(json.Unmarshal(cfgJSON, &cfgMap)).To(Succeed())
@@ -114,8 +114,8 @@ var _ = Describe("Task", Ordered, func() {
 		svcTask := task.Task{
 			Key:    task.NewKey(rack.NewKey(1, 1), 1),
 			Name:   "test-task",
-			Type:   runtime.TaskType,
-			Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+			Type:   arctask.TaskType,
+			Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 		}
 		return MustSucceed(factory.ConfigureTask(ctx, svcTask))
 	}
@@ -206,7 +206,7 @@ var _ = Describe("Task", Ordered, func() {
 
 	Describe("Factory.ConfigureTask", func() {
 		It("Should return ErrTaskNotHandled for non-arc task types", func(ctx SpecContext) {
-			factory := MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+			factory := MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 				Channel: channel.Wrap(dist.Channel),
 				Framer:  dist.Framer,
 				Status:  statusSvc,
@@ -231,7 +231,7 @@ var _ = Describe("Task", Ordered, func() {
 		})
 
 		It("Should return error for invalid config", func(ctx SpecContext) {
-			factory := MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+			factory := MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 				Channel:    channel.Wrap(dist.Channel),
 				Framer:     dist.Framer,
 				Status:     statusSvc,
@@ -239,7 +239,7 @@ var _ = Describe("Task", Ordered, func() {
 			}))
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 1),
-				Type:   runtime.TaskType,
+				Type:   arctask.TaskType,
 				Config: map[string]any{"arc_key": "not-a-valid-uuid"},
 			}
 			Expect(factory.ConfigureTask(ctx, svcTask)).Error().
@@ -247,7 +247,7 @@ var _ = Describe("Task", Ordered, func() {
 		})
 
 		It("Should return error when CompileProgram fails", func(ctx SpecContext) {
-			factory := MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+			factory := MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 				Channel:    channel.Wrap(dist.Channel),
 				Framer:     dist.Framer,
 				Status:     statusSvc,
@@ -255,15 +255,15 @@ var _ = Describe("Task", Ordered, func() {
 			}))
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 1),
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			Expect(factory.ConfigureTask(ctx, svcTask)).Error().
 				To(MatchError(query.ErrNotFound))
 		})
 
 		It("Should set error status when config is invalid", func(ctx SpecContext) {
-			factory := MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+			factory := MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 				Channel:    channel.Wrap(dist.Channel),
 				Framer:     dist.Framer,
 				Status:     statusSvc,
@@ -272,7 +272,7 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 2),
 				Name:   "test-invalid-config",
-				Type:   runtime.TaskType,
+				Type:   arctask.TaskType,
 				Config: map[string]any{"arc_key": "not-a-valid-uuid"},
 			}
 			Expect(factory.ConfigureTask(ctx, svcTask)).Error().
@@ -287,7 +287,7 @@ var _ = Describe("Task", Ordered, func() {
 		})
 
 		It("Should set error status when GetProgram fails", func(ctx SpecContext) {
-			factory := MustSucceed(runtime.NewFactory(runtime.FactoryConfig{
+			factory := MustSucceed(arctask.NewFactory(arctask.FactoryConfig{
 				Channel:    channel.Wrap(dist.Channel),
 				Framer:     dist.Framer,
 				Status:     statusSvc,
@@ -296,8 +296,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 3),
 				Name:   "test-module-not-found",
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			Expect(factory.ConfigureTask(ctx, svcTask)).Error().
 				To(MatchError(query.ErrNotFound))
@@ -320,8 +320,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 4),
 				Name:   "test-config-success",
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			t := MustSucceed(
 				newGraphFactory(simpleGraph(ch.Key())).
@@ -348,8 +348,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:  task.NewKey(rack.NewKey(1, 1), 5),
 				Name: "test-auto-start",
-				Type: runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{
+				Type: arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{
 					ArcKey:    uuid.New(),
 					AutoStart: true,
 				}),
@@ -428,8 +428,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 1),
 				Name:   "test-bad-node",
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			Expect(newGraphFactory(badNodeGraph).ConfigureTask(ctx, svcTask)).
 				Error().To(MatchError(ContainSubstring("undefined symbol")))
@@ -1265,8 +1265,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 100),
 				Name:   "test-div-zero",
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			t := MustSucceed(newTextFactory(ctx, prog).ConfigureTask(ctx, svcTask))
 			Expect(t.Exec(ctx, task.Command{Type: "start"})).To(Succeed())
@@ -1386,8 +1386,8 @@ var _ = Describe("Task", Ordered, func() {
 			svcTask := task.Task{
 				Key:    task.NewKey(rack.NewKey(1, 1), 101),
 				Name:   "test-div-recover",
-				Type:   runtime.TaskType,
-				Config: configToMap(runtime.TaskConfig{ArcKey: uuid.New()}),
+				Type:   arctask.TaskType,
+				Config: configToMap(arctask.TaskConfig{ArcKey: uuid.New()}),
 			}
 			t := MustSucceed(newTextFactory(ctx, prog).ConfigureTask(ctx, svcTask))
